Move direction from an Option<usize> to a u8 formatted enum. We aren't supporting an arity >= 2 so it's unnecessary to have a larger tree size and due to the risk of encoding problems between usize on architectures, we should be hard-setting a lower-risk type.
Also add sync_fork.sh to help automate fork synchronization with the GitHub CLI

+++ akd_core/src/verify/base.rs
@@ -74,7 +74,8 @@ pub fn verify_nonmembership(
     ]);
 
     for i in 0..ARITY {
-        lcp_real = lcp_real.get_longest_common_prefix(proof.longest_prefix_children[i].label);
+        lcp_real =
+            lcp_real.get_longest_common_prefix(proof.longest_prefix_children[i as usize].label);
     }
 
     if lcp_real == EMPTY_LABEL {
